// frontend/src/lib/api/generated/models/TicketItem.ts
export class TicketItem {
    'id': number;
    'name': string;
    'availableQuantity'?: number | null;
    'soldQuantity'?: number | null;
    'quantity': number;
    'description'?: string | null;
    'price'?: number | null;
    'expiredAt'?: Date | null;
    'type'?: TicketTypeCode | null;
    'status'?: TicketStatusCode | null;
    'salesStartAt'?: Date | null;
    'salesEndAt'?: Date | null;
    'deliveryMethod'?: TicketDeliveryMethodCode | null;
    'cancellationPolicyCode'?: TicketCancellationPolicyCode | null;
    'cancellationPolicyExtraDescription'?: string | null;
    'purchaseble'?: boolean | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "availableQuantity",
            "baseName": "available_quantity",
            "type": "number",
            "format": ""
        },
        {
            "name": "soldQuantity",
            "baseName": "sold_quantity",
            "type": "number",
            "format": ""
        },
        {
            "name": "quantity",
            "baseName": "quantity",
            "type": "number",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "price",
            "baseName": "price",
            "type": "number",
            "format": ""
        },
        {
            "name": "expiredAt",
            "baseName": "expired_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "TicketTypeCode",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "TicketStatusCode",
            "format": ""
        },
        {
            "name": "salesStartAt",
            "baseName": "sales_start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "salesEndAt",
            "baseName": "sales_end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "deliveryMethod",
            "baseName": "delivery_method",
            "type": "TicketDeliveryMethodCode",
            "format": ""
        },
        {
            "name": "cancellationPolicyCode",
            "baseName": "cancellation_policy_code",
            "type": "TicketCancellationPolicyCode",
            "format": ""
        },
        {
            "name": "cancellationPolicyExtraDescription",
            "baseName": "cancellation_policy_extra_description",
            "type": "string",
            "format": ""
        },
        {
            "name": "purchaseble",
            "baseName": "purchaseble",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return TicketItem.attributeTypeMap;
    }

    public constructor() {
    }
}